refactor: rename shared-args.internal.ts to _shared-args.ts
• Move shared-args.internal.ts to _shared-args.ts for better file organization

• Update imports in _shared-args.ts and all command files

• Update imports in commands/session.ts, commands/mcp.ts, commands/daily.ts, commands/monthly.ts, commands/blocks.ts
