+++ src/_shared-args.ts
@@ -1,6 +1,6 @@
 import type { Args } from 'gunshi';
-import type { CostMode, SortOrder } from './types.internal.ts';
-import { CostModes, dateSchema, SortOrders } from './types.internal.ts';
+import type { CostMode, SortOrder } from './_types.ts';
+import { CostModes, dateSchema, SortOrders } from './_types.ts';
 
 /**
  * Parses and validates a date argument in YYYYMMDD format
