+++ src/commands/monthly.ts
@@ -9,7 +9,7 @@ import {
 import { formatDateCompact, getDefaultClaudePath, loadMonthlyUsageData } from '../data-loader.ts';
 import { detectMismatches, printMismatchReport } from '../debug.ts';
 import { log, logger } from '../logger.ts';
-import { sharedCommandConfig } from '../shared-args.internal.ts';
+import { sharedCommandConfig } from '../_shared-args.ts';
 import { formatCurrency, formatModelsDisplay, formatNumber, pushBreakdownRows, ResponsiveTable } from '../utils.internal.ts';
 
 export const monthlyCommand = define({
